+++ packages/javascript/lib/node/file/getParentFilePath.ts
@@ -1,20 +1,6 @@
 import { toValidFilePath } from './toValidFilePath.js';
+import { posix, win32 } from 'path';
 
 export function getParentFilePath(filePath: string, isWindows?: boolean): string {
-	const listPathItem = filePath.split(isWindows ? '\\' : '/');
-	let currentPath;
-
-	listPathItem.pop();
-
-	if (listPathItem.length === 1) {
-		if (isWindows) {
-			currentPath = 'C:\\';
-		} else {
-			currentPath = '/';
-		}
-	} else {
-		currentPath = listPathItem.join(isWindows ? '\\' : '/');
-	}
-
-	return toValidFilePath(currentPath, isWindows);
+	return toValidFilePath(isWindows ? win32.dirname(filePath) : posix.dirname(filePath), isWindows);
 }