+++ packages/javascript/lib/node/file/joinFilePath.ts
@@ -2,9 +2,5 @@ import { posix, win32 } from 'path';
 import { toValidFilePath } from './toValidFilePath.js';
 
 export function joinFilePath(isWindows: boolean, ...paths: string[]): string {
-	if (isWindows) {
-		return toValidFilePath(win32.join(...paths), true);
-	}
-
-	return toValidFilePath(posix.join(...paths), false);
+	return toValidFilePath(isWindows ? win32.join(...paths) : posix.join(...paths), isWindows);
 }