+++ packages/javascript/CHANGELOG.md
@@ -5,6 +5,7 @@
 - **BREAKING CHANGES**: The `isWindows` argument is no longer used in `getFileExtension`.
 - `getFileExtension`: Performance improvements and cleanups
 - `getFileName`: Performance improvements and cleanups
+- `toValidFilePath`: Performance improvements and cleanups
 
 ## 1.10.4 (2025-11-25)