+++ packages/javascript/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `getFileName`: Performance improvements and cleanups
 - `toValidFilePath`: Performance improvements and cleanups
 - `getParentFilePath`: Performance improvements and cleanups
+- `joinFilePath`: Performance improvements and cleanups
 
 ## 1.10.4 (2025-11-25)